Restructuration du domaine média, réimplémentation du schéma et du plantUML
