+++ src/Persons/Schemas/PersonSchema.ts
@@ -1,5 +1,5 @@
 import {Document} from "mongoose";
-import { Media } from "../../Media/MediaSchema";
+import { Media } from "../../Media/Schemas/MediaSchema";
 import { Status } from "../../Moderation/Schemas/StatusSchema";
 
 export interface PersonSchema extends Document {
